Move files to corresponding new architecture
