+++ tugas-alpro/src/main/java/tugasalpro/managers/RuteManager.java
@@ -1,9 +1,10 @@
-package tugasalpro;
+package tugasalpro.managers;
 
-import java.io.IOException;
-import java.net.URISyntaxException;
 import java.util.List;
 import java.util.Optional;
+
+import tugasalpro.Repository;
+import tugasalpro.models.Rute;
 
 public class RuteManager {
     private Repository<Rute> repository;
