+++ tugas-alpro/src/main/java/tugasalpro/models/JalurRute.java
@@ -2,8 +2,6 @@ package tugasalpro.models;
 
 import java.util.ArrayList;
 
-import tugasalpro.*;
-
 public class JalurRute{
     private String kodeJalur;
     private Rute ruteJalur;
