+++ tugas-alpro/src/main/java/tugasalpro/models/Kereta.java
@@ -1,4 +1,7 @@
-package tugasalpro;
+package tugasalpro.models;
+
+import tugasalpro.Gerbong;
+
 public class Kereta {
     private String kodeKereta;
     private String namaKereta;
