+++ tugas-alpro/src/main/java/tugasalpro/models/KeretaRute.java
@@ -1,4 +1,4 @@
-package tugasalpro;
+package tugasalpro.models;
 
 
 import java.util.ArrayList;
