+++ tugas-alpro/src/main/java/tugasalpro/models/Rute.java
@@ -1,4 +1,7 @@
-package tugasalpro;
+package tugasalpro.models;
+
+import tugasalpro.*;
+
 public class Rute {
     private String kodeRute;
     private Kota kotaAsal;
